feat: use rate limit guard
