+++ src/telegram/rate-limit.ts
@@ -1,0 +1,25 @@
+import * as rateLimit from 'telegraf-ratelimit';
+import { Context } from './types';
+
+const limitConfig = {
+  window: 3000,
+  limit: 1,
+  onLimitExceeded: () => {},
+  keyGenerator: (ctx: Context) => {
+    const keys = [];
+    const fromId = ctx?.from?.id;
+    const chatId = ctx?.chat?.id;
+
+    if (fromId) {
+      keys.push(fromId);
+    }
+
+    if (chatId) {
+      keys.push(chatId);
+    }
+
+    return keys.join('-');
+  },
+};
+
+export const rateLimitMiddleware = rateLimit(limitConfig);
